+++ app/grails-app/conf/DataSource.groovy
@@ -1,3 +1,5 @@
+//import org.grails.plugin.hibernate.filter.HibernateFilterDomainConfiguration
+
 dataSource {
     pooled = true
     driverClassName = "org.h2.Driver"
@@ -5,25 +7,27 @@ dataSource {
     password = ""
 }
 
-hibernate { // to hibernate 4
-    default_schema = "public"
-    cache.use_second_level_cache = true
-    cache.use_query_cache = false  // LEGACY
-    //cache.region.factory_class = 'org.hibernate.cache.ehcache.SingletonEhCacheRegionFactory' // hibernate4: CAUTION: USE FOR DB-MIGRATION-PLUGIN
-    cache.region.factory_class = 'net.sf.ehcache.hibernate.SingletonEhCacheRegionFactory' // LEGACY: CAUTION: USE FOR DB-MIGRATION-PLUGIN
+hibernate {
+    default_schema                  = "public"
+    cache.use_second_level_cache    = true
+    cache.use_query_cache           = false // LEGACY
+    cache.region.factory_class      = 'org.hibernate.cache.ehcache.SingletonEhCacheRegionFactory' // hibernate4: CAUTION: USE FOR DB-MIGRATION-PLUGIN
+    //cache.region.factory_class    = 'net.sf.ehcache.hibernate.SingletonEhCacheRegionFactory' // LEGACY: CAUTION: USE FOR DB-MIGRATION-PLUGIN
+    //cache.region.factory_class    = 'org.hibernate.cache.ehcache.EhCacheRegionFactory' // Hibernate 4 .. this is NEW
 }
 
 // environment specific settings
 environments {
     development {
         dataSource {
-            dbCreate = "update"
+            dbCreate        = "update"
             driverClassName = "org.postgresql.Driver"
-            dialect = "org.hibernate.dialect.PostgreSQLDialect"
-            username = "laser"
-            password = "laser"
-            url = "jdbc:postgresql://localhost:5432/laser"
-            pooled = true
+            dialect         = "org.hibernate.dialect.PostgreSQLDialect"
+            username        = "laser"
+            password        = "laser"
+            url             = "jdbc:postgresql://localhost:5432/laser"
+            //configClass     = HibernateFilterDomainConfiguration.class
+            pooled          = true
             properties {
                 maxActive = -1
                 minEvictableIdleTimeMillis=1800000
@@ -38,13 +42,14 @@ environments {
     }
     test {
         dataSource {
-            dbCreate = "create-drop"
+            dbCreate        = "create-drop"
             driverClassName = "org.postgresql.Driver"
-            dialect = "org.hibernate.dialect.PostgreSQLDialect"
-            username = "laser"
-            password = "laser"
-            url = "jdbc:postgresql://localhost:5432/laser"
-            pooled = true
+            dialect         = "org.hibernate.dialect.PostgreSQLDialect"
+            username        = "laser"
+            password        = "laser"
+            url             = "jdbc:postgresql://localhost:5432/laser"
+            //configClass     = HibernateFilterDomainConfiguration.class
+            pooled          = true
             properties {
                 maxActive = -1
                 minEvictableIdleTimeMillis=1800000
@@ -59,12 +64,13 @@ environments {
     }
     production {
         dataSource {
-            dbCreate = "none" // disabled due database migration plugin; overwritten on dev-server
+            dbCreate        = "none" // disabled due database migration plugin; overwritten on dev-server
             driverClassName = "org.postgresql.Driver"
-            dialect = "org.hibernate.dialect.PostgreSQLDialect"
-            username = "laser"
-            password = "laser"
-            url = "jdbc:postgresql://localhost:5432/laser"
+            dialect         = "org.hibernate.dialect.PostgreSQLDialect"
+            username        = "laser"
+            password        = "laser"
+            url             = "jdbc:postgresql://localhost:5432/laser"
+            //configClass     = HibernateFilterDomainConfiguration.class
             properties {
                 maxActive = -1
                 minEvictableIdleTimeMillis=1800000