+++ app/src/groovy/de/laser/helper/DateUtil.groovy
@@ -16,4 +16,11 @@ class DateUtil {
         def format = messageSource.getMessage('default.date.format.notime', null, locale)
         new SimpleDateFormat(format)
     }
+
+    static DateFormat getSimpleDateFormat_NoZ(){
+        def messageSource = Holders.grailsApplication.mainContext.getBean('messageSource')
+        def locale = org.springframework.context.i18n.LocaleContextHolder.getLocale()
+        def format = messageSource.getMessage('default.date.format.noZ', null, locale)
+        new SimpleDateFormat(format)
+    }
 }