+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 
+0.14 PRE
+
+TODO
+ 
+- upgrade to hibernate 4.3.10
+- ehcache upgrade
+
+TODO
+
+
 0.13
 
 - reworked and improved finance handling
@@ -9,10 +19,18 @@
 - improved excel export
 - improved property management
 - improved namespace and identifier management
+- added yoda dashboard
 - improved profiler
 - added new event logging
+- added data consistency check
 - added various paginations
+- reworked dashboard tabs
+- improved ui dialogs and buttons
+- clearable filter dropdowns
+- fixed current titles date filter
+- various ui improvements
 - various bugfixes
+- upgrade to semantic-ui 2.4.2
 
 0.12.3